refactor table to new component
