+++ src/components/TeamTable/TeamTable.styles.js
@@ -1,0 +1,7 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+export const useStyles = makeStyles(theme => ({
+    table: {
+      marginTop: '16px'
+    }
+}));
