+++ src/pages/TeamPage/TeamPage.styles.js
@@ -6,13 +6,10 @@ export const useStyles = makeStyles(theme => ({
         backgroundColor: theme.palette.background.default
       }
     },
-    table: {
-      marginTop: '16px'
-    },
     header: {
       display: 'flex'
     },
     title: {
       paddingRight: '16px'
     }
-  }));
+}));
